feat: Add 'resources' prop to Refine component
Adds the 'resources' prop to the Refine component, allowing it to receive an array of resources. This change enables dynamic configuration of the resources to be used by Refine.

The 'resources' prop accepts an array of objects, where each object represents a resource and contains information such as name, endpoint, etc.